+++ firmware/esp32/main/comm/packet_router.h
@@ -53,6 +53,7 @@ class PacketRouter {
     void handle_browse_resp(const uint8_t* data, size_t length);
     void handle_status_response(const WaveX::Protocol::SampleMemStatusMessage& msg);
     void handle_sample_status(const WaveX::Protocol::SampleStatusMessage& msg);
+    void handle_storage_status(const WaveX::Protocol::StorageStatusMessage& msg);
     void handle_sample_stop_resp(const WaveX::Protocol::SampleStopRespMessage& msg);
     void handle_error(const WaveX::Protocol::ErrorMessage& msg);
     void handle_cv_cal_resp(const WaveX::Protocol::CvCalMessage& msg);